feat(rpc) Implement `Filecoin.EthGetLogs` (#4780)

+++ src/rpc/types/mod.rs
@@ -521,10 +521,10 @@ pub struct MiningBaseInfo {
 
 lotus_json_with_self!(MiningBaseInfo);
 
-#[derive(Clone, JsonSchema, Serialize, Deserialize)]
+#[derive(Debug, Clone, JsonSchema, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct EventEntry {
-    pub flags: u8,
+    pub flags: u64,
     pub key: String,
     pub codec: u64,
     pub value: LotusJson<Vec<u8>>,
